+++ lib/src/gitattributes.rs
@@ -13,7 +13,6 @@
 // limitations under the License.
 
 #![expect(missing_docs)]
-#![allow(unused)]
 
 use std::collections::HashMap;
 use std::collections::HashSet;