+++ lib/src/local_working_copy.rs
@@ -90,6 +90,12 @@ use crate::fsmonitor::FsmonitorSettings;
 use crate::fsmonitor::WatchmanConfig;
 #[cfg(feature = "watchman")]
 use crate::fsmonitor::watchman;
+#[cfg(feature = "git")]
+use crate::git::GitSettings;
+use crate::gitattributes::DiskFileLoader;
+use crate::gitattributes::GitAttributes;
+use crate::gitattributes::SearchPriority;
+use crate::gitattributes::TreeFileLoader;
 use crate::gitignore::GitIgnoreFile;
 use crate::lock::FileLock;
 use crate::matchers::DifferenceMatcher;
@@ -973,6 +979,10 @@ pub struct TreeStateSettings {
     pub exec_change_setting: ExecChangeSetting,
     /// The fsmonitor (e.g. Watchman) to use, if any.
     pub fsmonitor_settings: FsmonitorSettings,
+
+    /// Names of .gitattributes filters whose matching files should be ignored
+    /// in the working copy.
+    pub ignore_filters: HashSet<String>,
 }
 
 impl TreeStateSettings {
@@ -983,6 +993,19 @@ impl TreeStateSettings {
             eol_conversion_mode: EolConversionMode::try_from_settings(user_settings)?,
             exec_change_setting: user_settings.get("working-copy.exec-bit-change")?,
             fsmonitor_settings: FsmonitorSettings::from_settings(user_settings)?,
+            ignore_filters: {
+                #[cfg(feature = "git")]
+                {
+                    GitSettings::from_settings(user_settings)?
+                        .ignore_filters
+                        .into_iter()
+                        .collect()
+                }
+                #[cfg(not(feature = "git"))]
+                {
+                    HashSet::new()
+                }
+            },
         })
     }
 }
@@ -1007,6 +1030,10 @@ pub struct TreeState {
     exec_policy: ExecChangePolicy,
     fsmonitor_settings: FsmonitorSettings,
     target_eol_strategy: TargetEolStrategy,
+
+    // attributes
+    git_attributes: Arc<GitAttributes>,
+    ignore_filters: HashSet<String>,
 }
 
 #[derive(Debug, Error)]
@@ -1067,14 +1094,21 @@ impl TreeState {
             eol_conversion_mode,
             exec_change_setting,
             fsmonitor_settings,
+            ignore_filters,
         }: &TreeStateSettings,
     ) -> Self {
         let exec_policy = ExecChangePolicy::new(*exec_change_setting, &state_path);
+        let tree = store.empty_merged_tree();
+
+        let store_file_loader = TreeFileLoader::new(tree.clone());
+        let disk_file_loader = DiskFileLoader::new(working_copy_path.clone());
+
+        let git_attributes = Arc::new(GitAttributes::new(store_file_loader, disk_file_loader));
         Self {
             store: store.clone(),
             working_copy_path,
             state_path,
-            tree: store.empty_merged_tree(),
+            tree,
             file_states: FileStatesMap::new(),
             sparse_patterns: vec![RepoPathBuf::root()],
             own_mtime: MillisSinceEpoch(0),
@@ -1084,6 +1118,9 @@ impl TreeState {
             exec_policy,
             fsmonitor_settings: fsmonitor_settings.clone(),
             target_eol_strategy: TargetEolStrategy::new(*eol_conversion_mode),
+            // TODO We should update git_attributes every time TreeState::tree_id is updated
+            git_attributes,
+            ignore_filters: ignore_filters.clone(),
         }
     }
 
@@ -1332,6 +1369,8 @@ impl TreeState {
                 error: OnceLock::new(),
                 progress: *progress,
                 max_new_file_size: *max_new_file_size,
+                git_attributes: self.git_attributes.clone(),
+                ignore_filters: self.ignore_filters.clone(),
             };
             let directory_to_visit = DirectoryToVisit {
                 dir: RepoPathBuf::root(),
@@ -1503,6 +1542,9 @@ struct FileSnapshotter<'a> {
     error: OnceLock<SnapshotError>,
     progress: Option<&'a SnapshotProgress<'a>>,
     max_new_file_size: u64,
+
+    git_attributes: Arc<GitAttributes>,
+    ignore_filters: HashSet<String>,
 }
 
 impl FileSnapshotter<'_> {
@@ -1643,7 +1685,16 @@ impl FileSnapshotter<'_> {
             if let Some(progress) = self.progress {
                 progress(&path);
             }
-            if maybe_current_file_state.is_none()
+            if self
+                .git_attributes
+                .filter_matches(&path, &self.ignore_filters, SearchPriority::Disk)
+                .block_on()
+            {
+                // Skip gitattributes files that we want to ignore - this
+                // would result in them showing up as deleted, but we also
+                // omit them in `emit_deleted_files` to avoid that.
+                Ok(None)
+            } else if maybe_current_file_state.is_none()
                 && (git_ignore.matches_file(&path) && !self.force_tracking_matcher.matches(&path))
             {
                 // If it wasn't already tracked and it matches
@@ -1786,6 +1837,13 @@ impl FileSnapshotter<'_> {
             .flat_map(|(_, chunk)| chunk)
             // Whether or not the entry exists, submodule should be ignored
             .filter(|(_, state)| state.file_type != FileType::GitSubmodule)
+            // Whether or not the entry exists, ignored gitattributes files should be omitted
+            .filter(|(path, _)| {
+                !self
+                    .git_attributes
+                    .filter_matches(path, &self.ignore_filters, SearchPriority::Disk)
+                    .block_on()
+            })
             .filter(|(path, _)| self.matcher.matches(path))
             .try_for_each(|(path, _)| self.deleted_files_tx.send(path.to_owned()))
             .ok();